requirements.txt
asgiref==3.11.1
Django==6.0.2
django-allauth
python-dotenv
pillow
sqlparse==0.5.5
tzdata==2025.3
